decryption/pyproject.toml
[tool.poetry]
name = "encryption"
version = "0.1.0"
description = ""
authors = ["Jedrzej Szadejko <jendziu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.11.0"
pylint = "^2.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
